fix-camera spin
The camera would spin because when setting the position and rotation the bytes were offset by enough to offset the variables.
This was because the new operator was being overloaded, also the delete.
Also added a reset all inputs function so that the keys aren't up and active by default. This led to weird behaviour when it was up.